session-5-events-contract/apache-avro/requirements.txt
avro-python3==1.10.2